pyproject.toml
[project]
name = "contactdoc"
version = "0.1.0"
description = "AFDB ContactDoc pipeline: generate LLM training documents from AlphaFold DB structures"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gemmi>=0.6.0",
    "google-cloud-bigquery>=3.0",
    "google-cloud-storage>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
